pyproject.toml
[project]
name = "awesome-python"
version = "0.1.0"
description = "This project includes all the packages presented at the @thepypackager medium publications"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pyhumps>=3.8.0",
    "pylembic>=0.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["awesome-python"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.4",
]
